MRFL::hwc: fixed video playback never uses overlay again after failing to get rotated buffer from video driver.
BZ: 99458

Latest HWC impl only fallback to ST once failed to update the video layer & never
switch back to overlay even if rotated video buffer was ready.

This patch updated the error handling, so that hwc has a chance to switch back to
overlay when it detected the video layer can be updated to overlay successfully.

Change-Id: I53c5c208d06329e89464a08c0dca91200a91af4f
Signed-off-by: Jackie Li <yaodong.li@intel.com>
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Signed-off-by: Jackie Li <yaodong.li@intel.com>
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/101015
Reviewed-by: Hu, Austin <austin.hu@intel.com>
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>