Make HWC logging scalable.
BZ:100612

In current implemnention, if verbose logging is enabled there will be tons of messages printed out, which
makes it less useful for debugging purpose. We need one way to selectively enable or disable certain verbose
information.

For each log message, we also want class name and function name to be automatically pre-appended, this removes burden
of adding it manually, and enables developer to focus on key output message only.

Changes in this patch include:

1) define VTRACE, DTRACE, ITRACE, WTRACE and ETRACE, they are equivalent to LOGV, LOGD, LOGI, LOGW and LOGE
with class name and function name automatically pre-appended.
2) define CTRACE to trace function call entrance
3) define XTRACE to trace function call exit
4) CTRACE and XTRACE are turned off by default. They should only be enabled to trace deadlock, infinite loop or display freeze.
5) define ATRACE to trace function call entrance with input arguments, this is turned off by default either.

New type of trace can be defined to control output of certain type of information.

Change-Id: Iab185ad5ce407880322b2b306ea43bdb8440facc
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/101952
Reviewed-by: Hu, Austin <austin.hu@intel.com>
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>