Handle protected video playback.
BZ: 103301

This patch contains the following changes:

1) Define GraphicBuffer class and make it part of common library.
This class wraps buffer format and buffer usage as for now and provides
helper to check whether buffer is protected.

2) Avoid overhead of multiple data buffer creation during prepare, buffer is created
only once and attributes of buffer is stored internally for later use.

3) Ensure protected video is always rendered using harrdware overlay. IED needs to
be enabled when protected layer is received, a special hint is set in backbuffer
(OSTART_0Y) such that display controller can enable IED decryption during flip.

Change-Id: I32c4cf5f72e4fd650678264aaa9f0f63376f998b
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/102513
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Jin, Fu <fu.jin@intel.com>
Reviewed-by: Bish, Jim <jim.bish@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>