Presentation and video extended mode support for HDMI and WiDi
BZ: 103740

Implement  DisplayQuery class to check whether a layer is a video layer.

Implement DisplayAnalyzer to analyze contents first before layers are prepared,
if video extended mode is set by the analyzer then video layer in the primary layer is skipped.

If visible layer in video extended mode is 0 then display of primary
device can be turned off, otherwise, display of primary device needs
to be turned on always. (Turn off display will cause system to freeze,
so the feature is disabled for now).

Change-Id: I74711a7e8eab9a2055fba043a910e7db4fbfb1f9
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/104852
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>