hwc: fixed a defect during destroying overlay plane.
BZ: 111501

This fixed the tombstone issue happened during destroying overlay
plane.

The root cause of the tombstone is that deinitialize() of overlay
plane was called twice due to a logic error in code. this patch
fixed the error by deleting call to deinitialize() in OverlayPlaneBase
and use the deinitialize() in DisplayPlane only.

Note: this patch only fixed the code defect which would lead to
a tombstone, but we need dig in more to solve the real trigger of
hwcomposer destroying, which is MDS observer initialization error.

Change-Id: I43cb3e6b6202237250f0be9abd021ae6b3639acd
Signed-off-by: Jackie Li <yaodong.li@intel.com>
Signed-off-by: Austin Hu <austin.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/111724
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>