hwc: Start MDSObserver inner thread if fail to register the callback.
BZ: 111501

If SurfaceFlinger is crashed, the hwc module will be reload.

SurfaceFlinger and MultiDisplay services would restart, binder
transaction errors may occur when the MultiDisplayObserver going
to register the callback again.

So for this uncommnon case brought by the exceptions, the
MultiDisplayObserver needs to start its inner thread to do the
transactions with MDS.

Increased the life time for the thread since we meet large span
between the SF and MDS process start up.

Change-Id: I13c55e4b8077fe73e1f75353ac76a34f296a0ab5
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/112330
Reviewed-by: Zhu, Tianyang <tianyang.zhu@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>