hwc: fixed a mem leak issue
BZ: 114988

hwc had no chance to free data buffers when a reclaimed plane
was reused immediately for the next frame.

this patch fixed invalidation of the data buffers of a plane
an invalidation operation will NOT release the active data buffer
(which is being consumed by this plane). Meanwhile, this patch
also refined plane manager and made sure data buffers of primary
plane can be invalidated correctly.

Change-Id: I2c6d51de8322203640d287373c71cf00c39ceab4
Signed-off-by: Jackie Li <yaodong.li@intel.com>
Reviewed-on: http://android.intel.com:8080/114052
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: Patel, Hitesh K <hitesh.k.patel@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>