Fix the vpp regression by mapping kernel handle.
BZ: 126717

We haven't found a clear way to map the kernel handle so far.
Just had a workaround to do so through the Wsbm.
It's just needed when do the drm mode setting, so don't use
it as a formal way. Not sure about the root cause to make vpp
easily hang.
Need the DDK support for a reliable solution.

Change-Id: Ifff899f0afc624d8071db810726a040938777514
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/122581
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>