Reset layer composition type to HWC_FRAMEBUFFER if video state is change
BZ:125909

This bug is a corner case and application behaviors cause this error.
When the device is in lock state, using "am start" command
to start Gallery to play a streaming video, it is a strange operation,
then unlock, the Gallery is destoryed and created again,
after destroryed, some YUV layer still are sent to HWC via
SurfaceFlinger, it is very strange, related with Galeery behaviors and
this stange reproduced steps.
There is a confict with current HWC logic to handle video playback.
So, reset layers compostion type to "HWC_FRAMEBUFFER" when video state
is change, it is a common solution.

Change-Id: Ic9f13aa41a11cf4800279391e254f7ec608c1948
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/124389
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>