Fix widi video extended mode issue
BZ: 129562

WiDi is always in clone mode when playing video using gallery, as number
of video instances is updated only when HDMI is connected. If number of
video instances is 0, then video extended mode is disabled.

Fix is to update number of video instances regardless of HDMI connection status.

Changing one ITRACE to VTRACE as its coming per frame.

Change-Id: Id935ac15f127678d838a6ce7f1d110f949ab788a
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>
Reviewed-on: http://android.intel.com:8080/124829
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>