WIDI : Crop information needs to passed to widi.
BZ: 126897

If the widht & height does not align to 16 bytes then
there may be a green line in left side of the video during
extended mode.Crop information need to pass to widi to video
encoder if the clip contains rotate info.

This patch will pass the crop info from hwc to widi to video
encoder to solve this green bar issue on left side.

Signed-off-by: Ashish Singhi <ashish.singhi@intel.com>
Change-Id: If698d42c2a96b54344348637b13284fed321bc6d
Reviewed-on: http://android.intel.com:8080/126232
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>