Keep the HDMI mode invariable if app choose a specific mode.
BZ: 132903

The mode chosen by user may not match the current dynamic mode setting
requirement, because we cannot change the size of framebuffer in that
period.
So the mode will be unchanged after user set the mode in HDMI setting,
unless the cable is plug out or the DUT is reboot.

Change-Id: Id57879cb5903d5b07f40789607991451c7be322e
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/127824
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>