widi: Send frames to widi while seeking
BZ: 133768

This patch ensures that preview video frames are sent to widi
while a user drags the progress bar in media player applications.
This patch prevents black frames from being displayed on the remote
while seeking.

Since DisplayAnalyzer::checkVideoExtendedMode returns false while
seeking, we need to check each layer and verify that it is a
video layer but not a presentation layer. Once we find a frame that
meets the criteria we send it to widi.

Change-Id: Iad469c82f7eaa8c97deeeae4b7b371a1908dcdf1
Signed-off-by: Robert Crabtree <robert.crabtree@intel.com>
Reviewed-on: http://android.intel.com:8080/129642
Reviewed-by: Singhi, Ashish <ashish.singhi@intel.com>
Reviewed-by: Chowdhury, Rajneesh <rajneesh.chowdhury@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>