Reset the compositon type of all layers on trick mode.
BZ: 135120

HWC may attach some layers into HW to render. The Z-order needs to
be reconfigured if a video layer is attached to the overlay plane.
On trick mode, reset the compositon type of all layers to FB, thus
the hwc can analyze the layers again and correct the display Z-order.

Change-Id: I09028229054b6032b1a7623cdd2ab9a36221acbb
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/129690
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>