widi: Fix video size during rotation animation
BZ: 136431

VirtualDevice was calculating the frame size based on the
crop information from surface flinger. When the device
is rotated the crop information does not reflect the actual
video size. This causes HWC to send the wrong frame size
to widi which results in cropped video on the remote display.

This patch simply removes the frame size computation based
on the crop info and uses the frame size from the video metadata.

Change-Id: I23aba2d117a54ff43da53ec591b0bcfa10ee3e1f
Signed-off-by: Robert Crabtree <robert.crabtree@intel.com>
Reviewed-on: http://android.intel.com:8080/131464
Reviewed-by: Rogers, Brian E <brian.e.rogers@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>