widi: Optimize delivery time to widi
BZ: 146952

This patch will only get the video framerate from MDS when the
first video frame is delivered to HWC. Querying MDS takes a few
milliseconds which is too much on a per-frame basis.

This will reduce the number of frame drops in widi which will in
turn help with the video stutter issues we have been observing
during video playback.

Change-Id: I99a146661005381d4a8fb95fa23f2315ab84f399
Signed-off-by: Robert Crabtree <robert.crabtree@intel.com>
Reviewed-on: http://android.intel.com:8080/140119
Reviewed-by: Singhi, Ashish <ashish.singhi@intel.com>
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Tested-by: Post, DavidX J <davidx.j.post@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>