[HWC]: Fix snowflake on HDMI display
BZ: 146706

After video play, it will change the mode to prefer mode on HDMI display.
And when do mode setting, HWC will delete the old FB before.
This action will cause snowflake on HDMI display.
Because the new mode setting is not set well but the old fb is delete
when HDMI use the old mode to do flash.
So adjust the sequence in HWC when change mode. Delete the old fb after
set the new mode well.

Change-Id: Ic9926c7f411149b243427cc19246c0c9d7b37d7b
Signed-off-by: Shi Yang <yang.a.shi@intel.com>