[GFX-HWC] Wait and close acquire fence in HWC
BZ: 155916

For those layers whose compositionType are HWC_OVERLAY and framebufferTarget
layer VirtualDisplay, their acquire fence should be waited and
closed in HWC, or there will be an fd leak.

For the acquire fence of outbuf in VirtualDisplay, it should be waited
and closed in VirtualDisplay Application (WIDI), but in current
implementation hwc doesn't pass the fence to widi and we have to wait
and close it in HWC to prevent a fd leak problem.

Change-Id: I702bd59a0dcef9ce7156a1b27fb4f71e53ec824a
Signed-off-by: Lingyun Zhu <lingyun.zhu@intel.com>