[Workaround] Add a macro to distinguish MERR and MOFD.
BZ: 167139

Here we add a macro "MERR" to distinguish MERR and MOFD.
on MERR, we use "overlay semi-fixed" to solve overlay crash and MIPI off
issue, it make MERR stabler than before.
But this method is not available on MOFD, and cause this bug.
so we add this macro to as a workaround.
And we will continue to investigate this issue for root cause and remove
this patch in the future.

Change-Id: Ie15c25d194aea30169fe6e70b4fe821052d19767
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>