widi: Use onFramePrepare()
BZ: 183335

The onFramePrepare() method of IFrameListener allows hwcomposer to
tell WiDi that a frame will be arriving with a particular timestamp.
This allows WiDi to avoid generating a duplicate frame when a new
frame is about to be generated.

This patch enables use of onFramePrepare() when we're in clone mode,
since in this mode there is a time delta between knowing the frame is
going to be rendered and having the color-converted data ready to be
sent to WiDi.

Change-Id: Ib942d12eef48fb3ab9264329df183df2a8ccc73d
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>