widi: Fix for going into clone mode after shutdownVideo call
BZ: 177400

Once we call shutdownVideo(), existing code doesn't allow us to
recreate the video source ( using call to bufferInfoChanged() )
unless outputFrameInfo has changed. Modified the code so as to set
the forceNotifyBufferInfo flag to true after shutdownVideo() call,
so that we can recreate the video source in the clone mode as well.

Signed-off-by: Mukesh Arora <mukeshx.arora@intel.com>
Change-Id: Ie945e22818feba42453b35cfbe4543caa7f42324