hwc: fix hotplug thread can't exit immediately when deinit
BZ: 183366

Hotplug thread will poll netlink socket for hotplug event,
when deinit, this socket will be closed, but poll will never
return until there is a hotplug event(this is the stardard behavior).
Now, we add a pipe into the poll, when deinit we will close the
write end and read end will be signaled immediately.

Change-Id: Iaf1bf4164f5dd027c54d1e899f9ad1e9883591ed
Signed-off-by: Tan Wei <weix.z.tan@intel.com>