HWC: modify overlay ctx pipe to be same with other planes
BZ:184545

This is HWC patch to co-work with the kernel patch of plane
enable/disable auto control in driver. It modifies overlay
ctx pipe to be same with other planes. It also adds isDisabled
interface for HWC to query plane states in driver.

The kernel patch is 174233. Both patches needed to make it work.

Change-Id: I625bf7345b670241e816f948553d623058566171
Signed-off-by: Lei Zhang <lei.zhang@intel.com>