widi: Enable starting widi in extended mode
BZ: 179767

Earlier, we forced not to start the extended mode until
clone mode has started. Removed the code that was forcing
the widi to first start in clone mode and then go to extended
mode. Now, if there is a video layer availble ,we directly
start it in the extended mode.

Change-Id: If808a966a5eb9e0595c08fef99d1fd95c3ec2dd9
Signed-off-by: Mukesh Arora <mukeshx.arora@intel.com>