widi: Support non-composited protected video for AOSP Miracast
BZ: 191688

The TARGET_FORCE_HWC_FOR_VIRTUAL_DISPLAYS flag was added to
BoardConfig.mk to ensure that we have separate buffers for the target
frame buffer and outbuf, and HWC is responsible for copying image data
from the first to the second. A graphics patch ensures that outbuf is
created in the NV12 format.

This patch uses the above changes to have HWC handle the color
conversion for the AOSP Miracast path by doing a blit from the RGBA
frame buffer to the NV12 outbuf in commit. The outbuf content is what
gets fed to the encoder.

When a protected video layer is present, we stream just the video
without any compositing by copying the video data directly to outbuf.
The encoder has IED armed, so it can read the encrypted video data.

If the video is smaller than the frame buffer, we center it in outbuf
aligned with the x position on 16-byte boundary. This alignment is
necessary because the encryption is done with 16-byte blocks. We also
clear the border area around the video in this case, but since we
don't have encrypted black data to write into it, you will see
colorful stripes there.

1080p video will be scrambled currently, because handling of the tiled
format and downscaling aren't present. A later patch will add this,
and also composition support using VSP.

This new code is active when the FrameServer interface that Intel WiDi
uses is not in use. When the FrameServer interface is used, the
previous behaviors are maintained.

Change-Id: Ia295a71cb0436611e427edbd8ebc6b6c146af6de
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>