hwc: fix last frame flashing on MIPI under ext video mode
BZ:193720

When none FB layers, it does not need to do plane allocation.
This happens when entering ext video mode, or receiving call
with HDMI connected, where layers will be marked as SKIPPED.

Root cause for MIPI DBI FIFO not empty issue after apply this
patch is found and fixed in driver patch 192143.

Change-Id: I4d0601664783e8730782856c5da7db024a8a273c
Signed-off-by: Lei Zhang <lei.zhang@intel.com>