widi: Don't keep cached buffers in RGBA/BGRA conversion case
BZ: 199694

We keep a cache of mapped buffers and clear it when we no longer need
the mappings. In the case triggered by the VirtualDisplay testcases in
GTS, we map the framebuffer target and outbuf so we can do RGBA/BGRA
conversion on the CPU. We unmap these buffers when the virtual display
becomes blank. In other words, on the first frame rendered after the
VirtualDisplaySurface is destroyed.

Unmapping the handles at this point is triggering a double-free that
corrupts the heap and crashes SurfaceFlinger. To work around this, do
the color conversion synchronously and delete the cached mappings
before returning to SurfaceFlinger, so that VirtualDisplaySurface
can't be destroyed before we unmap.

There is a remaining SurfaceFlinger crash that isn't from heap
corruption, so this patch does not ensure 100% stability, but this
does solve the more common crash that is a regression of the patch
'Support secure video playback with VSP composition'.

Change-Id: I130383840db771bef56b4a67c8883e9d9ba982d6
Signed-off-by: Brian Rogers <brian.e.rogers@intel.com>