widi: Fix crash in hwcomposer during long video playback
BZ: 200546

There was a synchronization issue during map and unmap in
BufferManager.cpp. When CachedBuffer is destroyed, it unmaps the mapper.
Before unmap() completes, a new CachedBuffer is created and executes map().
getMapper returns the same mapper from mBufferPool as it is still not unmaped.
After this, the unmap() function unmaps this mapper.  mPayloadManager->getMetaData()
uses this mapper and this results in a crash.

Crash was not seen after synchronizing the map() and unmap() functions.

Change-Id: Ibb0339aa90c167f04336c23c72eb104747e1644b
Signed-off-by: Sandeep Ramankutty <sandeepx.ramankutty@intel.com>