HWC: reset vsync source eariler
BZ: 204352

when hdmi is plugged out, vsync source will be switched to PIPE A.
however if the vsync thread is requesting vblank on PIPE B, it will
return error. This will cause the thread run like a infinite loop. the
fix is switching vsync source eariler and sleep 16ms if drmWaitvblank
returns error from kernel, this will give chance for other thread to
run.

Change-Id: I2ef8084f062acd043307954473d5a1a5620be689
Signed-off-by: Geng, Xiujun <xiujun.geng@intel.com>