[Workaround] HWC: avoid using overlay A on pipe B
BZ: 204776

Based on spec, overlay switch from one pipe to another takes 2 vsync to complete.
When playing transformed vdieo (overlay C can't be used as there is rotation) and
video extended mode is entered (HDMI is connected) overlay A will be assigned
to pipe B from pipe A. But as vsync is disabled on MIPI overlay swich is never
completed.

Workaround is not to assign overlay A on pipe B.

Change-Id: Id3e0a3e8717d95503154730dbc35dfb05ad18acb
Signed-off-by: ljia5 <lin.a.jia@intel.com>
Signed-off-by: Andy Qiu <junhai.qiu@intel.com>